third_party/blink/web_tests/platform/generic/fast/css/pseudo-valid-007-expected.txt

This test performs a check for the :valid CSS selector on a select element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Lorem ipsum
 
empty
another
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS document.defaultView.getComputedStyle(v[i], null).getPropertyValue('background-color') is 'rgb(0, 255, 0)'
PASS successfullyParsed is true

TEST COMPLETE
